// final-project-venturousman/backend/src/models/itineraryViewModel.ts
interface ItineraryViewModel {
    id: string;
    destination: string;
    longitude: number;
    latitude: number;
    startDate: string;
    endDate: string;
    preferences: string;
    userId: string;
    recommendationId: string;
    shareableId: string;
    title: string;
    content: string;
}

export default ItineraryViewModel;